Ensure bbs healthcheck server has shut down in every AfterEach
* Moved killing of bbsProcess to main_suite_test.go so we can shut down
bbs process the same way everywhere

* Because the BBS healthcheck server is the first thing to come up and
the last thing to be killed when the process is taken down, we saw
frequent races between killing the BBS and starting it up in the next
test. Now, we wait for the healthcheck to be unavailable before moving
on to the next test.

Signed-off-by: Jen Spinney <jen.spinney@hpe.com>