Ensure that all functions in sqldb can operate with one database connection
this fix a deadlock that can occur if more than one method in the package
requires more than one database connection to complete, e.g. assuming
CreateDesiredLRP and StartActualLRP had the following code:

  1. begin transaction
  2. run a query outside the transaction using a new connection
  3. end transaction

if the maximum open connections is set to 2 we can end up in a deadlock
situtation where both functions are called in separated goroutines and both of
them grab a connection to being the transaction but will fail to proceed with
the second step, since golang sql package will prevent new connections from
being created.

[#136754083]