BBS should retire the proper lrp index instead of using the loop counter
the loop counter will break if there are gaps in the lrp indexes, e.g. if an
app is scaled down from 4 instances to 1 instance but only indexes 1 & 2 were
stopped but 0 & 3 remain running. In that case the bbs will continue trying to
retire an lrp with index `1' and fail because the lrp doesn't exist anymore

[#144050159]