WIP: add new events endpoint using v2 desiredLRP format
This is functionally complete, but there's a super gross interface
(hadnlers.EventController) which is only there for testing and is
just... wrong!

Unfortunately the current structure of the event handler tests make it
difficult to remove this wihtout duplicating a lot of tests. It probably
makes sense to take a step back and revisit the structure of those tests
from the top down.

[#155728323]