wip: started writing integration tests
I started implementing the logic after Row 5 and felt that the responsibiliity
is going to be split between the controller and the database in non-trivial way
which required some higher level integration tests. I'm also worried about
breaking the implementation when we start the Flattened actual LRP track
without proper integration tests.