Fix a flaky test in BBS
The 'StartActualLRP' used to have a racy side effect when the convergence tried
to simultaneously call CreateUnclaimedLRP. The fix is to wait for the auction
to take place which guarantees the unclaimed LRP has been created.

We should replace this with an
Eventually(client.GetFlatLRPs).Should(ContainUnclaimed) once the flat LRP api
is in place.

[#158123373](https://www.pivotaltracker.com/story/show/158123373)