Sometimes emit ActualLRPCreatedEvent for replacement LRPs
In general, we don't want to emit CreatedEvents for LRPs created to
replace Suspect LRPs. However, if the suspect instance crashes and gets
removed, we then want to emit a CreatedEvent for replacement LRP
(both for Claimed or Unclaimed) because that ActualLRP is now the
active 'instance' for the ActualLRPGroup.

[#159597788](https://www.pivotaltracker.com/story/show/159597788)

Signed-off-by: Edwin Xie <exie@pivotal.io>