Do not log confusing errors when replacement instance becomes suspect too
Previously the BBS logged confusing errors:
  - change-actual-lrp-presence.failed-updating-lrp: "duplicate entry, key exists"
  - "failed-unclaiming-lrp"

The first error shows up because the suspect LRP for the LRP key already
exists, so transitioning the replacement to suspect will fail.

The second error shows up because the helper to mark the LRP as suspect
will unclaim the LRP if it fails to transition the replacement to
suspect. Then the convergence loop tries to create an unclaimed LRP
anyways (since it assumes it was handling the case where it was marking
the LRP suspect for the first time).

Instead of going through this convoluted workflow, we check during
convergence whether or not the suspect LRP already exists (is the LRP
for which its cell went missing a replacement?). If so, we just
transition it straight to unclaimed. Otherwise, we mark the LRP as
suspect and create an unclaimed LRP as usual.

[#159597788](https://www.pivotaltracker.com/story/show/159597788)

Signed-off-by: Edwin Xie <exie@pivotal.io>