Revamped evacuation handler
Use flat ActualLRP database apis instead of the LRPGroup ones.  Also revisit
the test suite to ensure they are correct.  We found that a lot of the tests
for the function seemed incorrect -- for example, most tests were assuming that
an normal instance and an evacuating instance were on the same cell, which we
is not possible.

[#160355261](https://www.pivotaltracker.com/story/show/160355261)

Signed-off-by: Nima Kaviani <nkavian@us.ibm.com>