Fix the test assertions
This test just started failing because of commit 574582f.  I think the behavior
and the test assertion originally matched.  That is because we originally
preferred Claimed/Ordinary over Claimed/Suspect.  At some point we decided to
change the priority in the resolution logic to prefer Claimed/Suspect over
Claimed/Ordinary and I think we didn't update this test or the implementation.