Handle nil/UNCLAIMED cases properly in event generation
We believe that the ActualLRPInstanceChangedEvent structure assumes that
any single ActualLRP should only ever contain the state for a single LRP
that gets scheduled onto some cell and that it wouldn't be reused for
rescheduling onto another cell. This means that if the LRP is suspect or being
evacuated, its replacement should be a new instance of an ActualLRP.

With that assumption in mind, it makes sense to dedupe the ActualLRPKey
and ActualLRPInstance key inside the changed event, since we shouldn't
get two ActualLRPs in the event with completely different Keys or
InstanceKeys.

However, we still need to handle edge cases like a changed event
from UNCLAIMED -> CLAIMED (which is expected) or RUNNING -> nil.

Also, we added some additional nil checks to make our unit tests happy.

[#159596156](https://www.pivotaltracker.com/story/show/159596156)

Co-authored-by: Edwin Xie <exie@pivotal.io>