Pass request context down to db calls
Removing CancellableContextRequest which was spinning up the goroutine
and running a request handler in it. In case if context is cancelled
this resulted in a goroutine still running and potentially modifying the
request/response. When http handler finishes go server handler continues
finishing the response. This can result in panic due to concurrent reads
and writes on the same response object (in particular headers map).

Instead pass the request down to db and let golang db drivers to handle
them appropriately.

[#165795047](https://www.pivotaltracker.com/story/show/165795047)

Co-authored-by: Edwin Xie <exie@pivotal.io>