Add routable field to actual_lrp
In order to be able to distinguish between actual_lrps that are ready to
serve traffic (i.e. have passed their readiness checks if present), we
need to add a field to the actual_lrps to persist this information.

We have decided to call this field `routable`, it is a boolean (can only
be true/false) and has a default value of `false` when the ActualLRP is
created.

This commit does the following:
- adds the `routable` field to the `actual_lrp` protobuf message
- regenerate the protobuf files
- adds `routable` field to the ActualLRP model
- automatically adds `routable: false` when creating an Unclaimed Actual
  LRP

We also updated the in-repo documentation to account for this change.

This commit is part of a larger effort to improve the app healthchecks.
To see what we are doing, look at this (publically available) design
doc:
- https://docs.google.com/document/d/1AS8GzKg3pdAv56nySQTgF4NaqqzrBwmztkL2_8ahKa4/edit#

[#185037071](https://www.pivotaltracker.com/story/show/185037071)

Signed-off-by: Renee Chu <reneec@vmware.com>