Avoid race condition when replacement for suspect LRP has started
BBS convergence is prone to race condition since it collects keys that
may become stale during convergence cycle.

When Suspect cell comes back up and SuspectKeysWithExistingCells gets
populated some suspect LRPs might have replacement running. When
replacement LRP starts StartActualLRP will remove suspect LRP from
database.

Previously, we tried to remove Ordinary LRP replacement for the Suspect
one first (which succeded). And then we tried to change Suspect LRP
presence to Ordinary and if that Suspect was removed already by
StartActualLRP for Ordinary we were getting an error changing presense.
This resulted in no LRPs registered for that key. Another issue was that
we didn't emit any events when we removed Ordinary LRP.

Now, we will remove ordinary LRPs only if there is a suspect LRP with
the same key and we check this in the same transaction.