UnavailableHandler takes multiple channels to specify service readiness
Thus far, the UnavailableHandler has returned 503's only for a single readiness check
(specifically, checking that migrations have all run). We're planning to add functionality
to check that the locket lock has been acquired as well. In preparation for that change,
we allow this handler to make checks against multiple readiness channels

Signed-off-by: David Sabeti <david.sabeti@broadcom.com>
Signed-off-by: Maria Shaldybin <maria.shaldybin@broadcom.com>